rename to citation.vim
